+++ srcs/vector.c
@@ -6,13 +6,13 @@
 /*   By: hthomas <hthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/28 13:49:41 by hthomas           #+#    #+#             */
-/*   Updated: 2020/04/19 12:07:47 by hthomas          ###   ########.fr       */
+/*   Updated: 2020/11/12 13:04:58 by hthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minirt.h"
 
-t_vect		new_vect(const float x, const float y, const float z)
+t_vect		new_vect(const double x, const double y, const double z)
 {
 	t_vect	new;
 
@@ -22,24 +22,24 @@ t_vect		new_vect(const float x, const float y, const float z)
 	return (new);
 }
 
-float	distance(const t_vect p1, const t_vect p2)
+double	distance(const t_vect p1, const t_vect p2)
 {
 	return (sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2) + pow(p2.z - p1.z, 2)));
 }
 
 t_vect		normalize(const t_vect vect)
 {
-	float	length;
+	double	length;
 
 	length = distance(new_vect(0,0,0), vect);
 	return (new_vect(vect.x / length, vect.y / length, vect.z / length));
 }
 
-t_vect		mult_mat(const float mat[3][3], const t_vect vect)
+t_vect		mult_mat(const double mat[3][3], const t_vect vect)
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 	t_vect	res;
 
 	x = mat[0][0] * vect.x + mat[0][1] * vect.x + mat[0][2] * vect.x;
@@ -49,17 +49,17 @@ t_vect		mult_mat(const float mat[3][3], const t_vect vect)
 	return (res);
 }
 
-t_vect		rot_vect(const t_vect vect, const float angle, const char axe)
+t_vect		rot_vect(const t_vect vect, const double angle, const char axe)
 {
-	float	alpha;
+	double	alpha;
 
 	alpha = to_rad(angle * 0.5);
 	if (axe == 'x')
-		return (mult_mat((float[3][3]){{1, 0, 0}, {0, cos(alpha), -sin(alpha)}, {0, sin(alpha), -cos(alpha)}}, vect));
+		return (mult_mat((double[3][3]){{1, 0, 0}, {0, cos(alpha), -sin(alpha)}, {0, sin(alpha), -cos(alpha)}}, vect));
 	else if (axe == 'y')
-		return (mult_mat((float[3][3]){{cos(alpha), 0, sin(alpha)}, {0, 1, 0}, {-sin(alpha), 0, cos(alpha)}}, vect));
+		return (mult_mat((double[3][3]){{cos(alpha), 0, sin(alpha)}, {0, 1, 0}, {-sin(alpha), 0, cos(alpha)}}, vect));
 	else if (axe == 'z')
-		return (mult_mat((float[3][3]){{cos(alpha), sin(alpha), 0}, {sin(alpha), cos(alpha), 0}, {0, 0, 1}}, vect));
+		return (mult_mat((double[3][3]){{cos(alpha), sin(alpha), 0}, {sin(alpha), cos(alpha), 0}, {0, 0, 1}}, vect));
 	return (new_vect(0, 0, 0));
 }
 
@@ -78,12 +78,12 @@ t_vect		minus_vect(const t_vect vect)
 	return (new_vect(-vect.x, -vect.y, -vect.z));
 }
 
-t_vect		multi_vect(const t_vect vect, const float x)
+t_vect		multi_vect(const t_vect vect, const double x)
 {
 	return (new_vect(vect.x * x, vect.y * x, vect.z * x));
 }
 
-float		dot_product(const t_vect vect1, const t_vect vect2)
+double		dot_product(const t_vect vect1, const t_vect vect2)
 {
 	return (vect1.x * vect2.x + vect1.y * vect2.y + vect1.z *vect2.z);
 }
@@ -92,8 +92,8 @@ t_vect		cross_product(const t_vect v1, const t_vect v2)
 {
 	t_vect	origin;
 	t_vect	normal;
-	float	denom;
-	float	cross;
+	double	denom;
+	double	cross;
 
 	origin = new_vect(0, 0, 0);
 	denom = distance(origin, v1) * distance(origin, v2);
